perf(core): decouple luma from Device defaults
