+++ modules/core/src/adapter/device.ts
@@ -39,6 +39,9 @@ import {textureFormatDecoder} from '../shadertypes/texture-types/texture-format-
 import type {ExternalImage} from '../shadertypes/image-types/image-types';
 import {isExternalImage, getExternalImageSize} from '../shadertypes/image-types/image-types';
 import {getTextureFormatTable} from '../shadertypes/texture-types/texture-format-table';
+import {DEVICE_DEFAULT_PROPS} from './device-defaults';
+
+export {_getDefaultDebugValue} from './device-defaults';
 
 /**
  * Identifies the GPU vendor and driver.
@@ -487,57 +490,7 @@ export interface DeviceFactory {
  * WebGPU Device/WebGL context abstraction
  */
 export abstract class Device {
-  static defaultProps: Required<DeviceProps> = {
-    id: null!,
-    powerPreference: 'high-performance',
-    failIfMajorPerformanceCaveat: false,
-    featureLevel: undefined!,
-    createCanvasContext: undefined!,
-    // WebGL specific
-    webgl: {},
-
-    // Callbacks
-    // eslint-disable-next-line handle-callback-err
-    onError: (error: Error, context: unknown) => {},
-    onResize: (context: CanvasContext, info: {oldPixelSize: [number, number]}) => {
-      const [width, height] = context.getDevicePixelSize();
-      log.log(1, `${context} resized => ${width}x${height}px`)();
-    },
-    onPositionChange: (context: CanvasContext, info: {oldPosition: [number, number]}) => {
-      const [left, top] = context.getPosition();
-      log.log(1, `${context} repositioned => ${left},${top}`)();
-    },
-    onVisibilityChange: (context: CanvasContext) =>
-      log.log(1, `${context} Visibility changed ${context.isVisible}`)(),
-    onDevicePixelRatioChange: (context: CanvasContext, info: {oldRatio: number}) =>
-      log.log(1, `${context} DPR changed ${info.oldRatio} => ${context.devicePixelRatio}`)(),
-
-    // Debug flags
-    debug: getDefaultDebugValue(),
-    debugGPUTime: false,
-    debugShaders: log.get('debug-shaders') || undefined!,
-    debugFramebuffers: Boolean(log.get('debug-framebuffers')),
-    debugFactories: Boolean(log.get('debug-factories')),
-    debugWebGL: Boolean(log.get('debug-webgl')),
-    debugSpectorJS: undefined!, // Note: log setting is queried by the spector.js code
-    debugSpectorJSUrl: undefined!,
-
-    // Experimental
-    _reuseDevices: false,
-    _cacheShaders: true,
-    _destroyShaders: false,
-    _cachePipelines: true,
-    _sharePipelines: true,
-    _destroyPipelines: false,
-    // TODO - Change these after confirming things work as expected
-    _initializeFeatures: true,
-    _disabledFeatures: {
-      'compilation-status-async-webgl': true
-    },
-
-    // INTERNAL
-    _handle: undefined!
-  };
+  static defaultProps: Required<DeviceProps> = {...DEVICE_DEFAULT_PROPS};
 
   get [Symbol.toStringTag](): string {
     return 'Device';
@@ -1098,37 +1051,4 @@ or create a device with the 'debug: true' prop.`;
 
     return newProps;
   }
-}
-
-/**
- * Internal helper for resolving the default `debug` prop.
- * Precedence is: explicit log debug value first, then `NODE_ENV`, then `false`.
- */
-export function _getDefaultDebugValue(logDebugValue: unknown, nodeEnv?: string): boolean {
-  if (logDebugValue !== undefined && logDebugValue !== null) {
-    return Boolean(logDebugValue);
-  }
-
-  if (nodeEnv !== undefined) {
-    return nodeEnv !== 'production';
-  }
-
-  return false;
-}
-
-function getDefaultDebugValue(): boolean {
-  return _getDefaultDebugValue(log.get('debug'), getNodeEnv());
-}
-
-function getNodeEnv(): string | undefined {
-  const processObject = (
-    globalThis as typeof globalThis & {
-      process?: {env?: Record<string, string | undefined>};
-    }
-  ).process;
-  if (!processObject?.env) {
-    return undefined;
-  }
-
-  return processObject.env['NODE_ENV'];
 }
